backend/requirements.txt
fastapi==0.110.0
uvicorn==0.24.0
sqlalchemy==2.0.25
pydantic==2.6.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
pydantic[email]
alembic==1.12.1
boto3==1.34.11
python-dotenv==1.0.0
pydantic-settings==2.0.3
